requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
decorator==4.4.2
future==0.18.2
nltk==3.4.5
numpy==1.18.2
pandas==1.0.3
pycparser==2.20
python-dateutil==2.8.1
python-telegram-bot==12.4.2
pytz==2019.3
six==1.14.0
soupsieve==2.0
tornado==6.0.4
urllib3==1.25.8
